Import unison-2.40.102nb22 as net/unison2.40 before updating net/unison.
Unison is a file-synchronization tool for Unix and Windows. It allows
two replicas of a collection of files and directories to be stored
on different hosts (or different disks on the same host), modified
separately, and then brought up to date by propagating the changes
in each replica to the other.