Import libretro-vba-next-20150105 as emulators/libretro-vba-next.
Libretro is a simple but powerful development interface that allows for the
easy creation of emulators, games and multimedia applications that can plug
straight into any libretro-compatible frontend. This development interface is
open to others so that they can run these pluggable emulator and game cores
also in their own programs or devices.

VBA Next is an optimized port of the VBA-M Nintendo Game Boy Advance video
game system emulator.