Update to 1.4.1. Should we remove this from pkgsrc? (see paragraph below)
A few cleanup patches that have been sitting on the branch for almost a
year. The void driver is a dummy driver that does nothing and just fulfils
the requirements of really old servers that must have an input device.
It is just dead weight for any server released after 2007 or so. Don't use
it.

Alan Coopersmith (4):
      Link with $(XORG_LIBS) to support no-undefined linking
      Fix an integer truncation warning from clang
      configure: Drop AM_MAINTAINER_MODE
      autogen.sh: Honor NOCONFIGURE=1

Gaetan Nadon (1):
      config: replace deprecated use of AC_OUTPUT with AC_CONFIG_FILES

Peter Hutterer (1):
      xf86-input-void 1.4.1

Terry Lambert (1):
      Return proper default for unknown values in pInfo->device_control.