Update to 0.27:
This version primarily fixes bugs, but it includes a few new features,
and hopefully will provide a solid base for 0.28, which may finally
include rm/gc support.

Of note since 0.26:

  - Files can be restored sparsely via "restore --sparse".  See
    bup-sparse(1) for the details.  Given that this option is
    completely new and affects the data written, some
    caution/verification may be wise (i.e. there were bugs in some of
    the development revisions that could mangle the output).

  - Duplicity backups can be imported via the *experimental* command
    "bup import-duplicity".  Please exercise caution, and let us know
    if you have any trouble.

  - Tests can be run in parallel now via "make -j check".

As before, the metadata support is somewhat immature, but it seems to
be holding up fairly well.  Though we still need to add better support
for cross-filesystem-type save/restore (which can be too noisy), etc.